requirements.txt
fastapi
uvicorn
jinja2
yfinance
sqlalchemy
google-cloud-bigquery
pandas
python-coinmarketcap
coinmarketcap
requests
python-dotenv
aiofiles